moved basicInit to a helper file
