+++ tests/auth.spec.ts
@@ -1,103 +1,5 @@
 import { test, expect } from 'playwright-test-coverage';
-import { Page } from 'playwright';
-import { User, Role } from '../src/service/pizzaService';
-
-async function basicInit(page: Page) {
-    let loggedInUser: User | undefined;
-    const validUsers: Record<string, User> = {
-        'd@jwt.com': {
-            id: '3',
-            name: 'Pizza Diner',
-            email: 'd@jwt.com',
-            password: 'a',
-            roles: [{ role: Role.Diner }],
-        },
-    };
-
-    // Authorize login for the given user
-    await page.route('*/**/api/auth', async (route) => {
-        const loginReq = route.request().postDataJSON();
-        if (route.request().method() === 'DELETE') {
-            loggedInUser = undefined;
-            await route.fulfill({ json: { message: 'logged out successfully' } });
-            return;
-        }
-        const user = validUsers[loginReq.email];
-        if (!user || user.password !== loginReq.password) {
-            await route.fulfill({ status: 401, json: { error: 'Unauthorized' } });
-            return;
-        }
-        loggedInUser = validUsers[loginReq.email];
-        const loginRes = {
-            user: loggedInUser,
-            token: 'abcdef',
-        };
-        expect(route.request().method()).toBe('PUT');
-        await route.fulfill({ json: loginRes });
-    });
-
-    // Return the currently logged in user
-    await page.route('*/**/api/user/me', async (route) => {
-        expect(route.request().method()).toBe('GET');
-        await route.fulfill({ json: loggedInUser });
-    });
-
-    // A standard menu
-    await page.route('*/**/api/order/menu', async (route) => {
-        const menuRes = [
-            {
-                id: 1,
-                title: 'Veggie',
-                image: 'pizza1.png',
-                price: 0.0038,
-                description: 'A garden of delight',
-            },
-            {
-                id: 2,
-                title: 'Pepperoni',
-                image: 'pizza2.png',
-                price: 0.0042,
-                description: 'Spicy treat',
-            },
-        ];
-        expect(route.request().method()).toBe('GET');
-        await route.fulfill({ json: menuRes });
-    });
-
-    // Standard franchises and stores
-    await page.route(/\/api\/franchise(\?.*)?$/, async (route) => {
-        const franchiseRes = {
-            franchises: [
-                {
-                    id: 2,
-                    name: 'LotaPizza',
-                    stores: [
-                        { id: 4, name: 'Lehi' },
-                        { id: 5, name: 'Springville' },
-                        { id: 6, name: 'American Fork' },
-                    ],
-                },
-                { id: 3, name: 'PizzaCorp', stores: [{ id: 7, name: 'Spanish Fork' }] },
-                { id: 4, name: 'topSpot', stores: [] },
-            ],
-        };
-        expect(route.request().method()).toBe('GET');
-        await route.fulfill({ json: franchiseRes });
-    });
-
-    // Order a pizza.
-    await page.route('*/**/api/order', async (route) => {
-        const orderReq = route.request().postDataJSON();
-        const orderRes = {
-            order: { ...orderReq, id: 23 },
-            jwt: 'eyJpYXQ',
-        };
-        expect(route.request().method()).toBe('POST');
-        await route.fulfill({ json: orderRes });
-    });
-
-    await page.goto('/');
-}
+import { basicInit } from './helpers';
 
 test('login and logout', async ({ page }) => {
     await basicInit(page);
